requirements.txt
aiogram==3.4.1
python-dotenv==1.0.1
httpx==0.27.2
pydantic==2.5.3
pypdf==5.0.1
SQLAlchemy==2.0.36
aiosqlite==0.20.0
uvloop==0.21.0; sys_platform != 'win32'
Pillow==10.4.0
faster-whisper>=1.0.3
requests>=2.31.0
openpyxl==3.1.2

